Run all the integration tests - 2024-09-13 PT

+++ slack-api-model/src/main/java/com/slack/api/model/block/element/RichTextSectionElement.java
@@ -54,7 +54,7 @@ public class RichTextSectionElement extends BlockElement implements RichTextElem
     public static class Date implements RichTextElement {
         public static final String TYPE = "date";
         private final String type = TYPE;
-        private Long timestamp;
+        private Integer timestamp;
         private String format;
         private TextStyle style;
         private String url;
